backend/requirements.txt

yfinance
requests
pandas
transformers
jinja2
supabase
python-dotenv
